Update README.md
For this Module 2 challenge, I worked with classmates and a tutor. The project was completed using Microsoft VBA. By working with multiple study groups, we troubleshooted each other's code and learned how we each came up with our code. I met with a tutor twice for this challenge, who helped revise my code, as well as lead me on areas where I struggled for both the practice file (alphabetical testing) and the actual homework file (Multiple year stock data). 

My tutor was able to help me with the following: 
Create a for loop to run my code through all sheets 
Conditional formatting to change the cell's color to red/green 
Conditional formatting to update the number format to "0.00%" 
Revised my code to calculate the greatest % increase, decrease, and total volume